// src - SiteVente avc DB Staticdans un restAPIService/app/models/product.model.ts
export class ProductModel {
  id?: any ;
  name?: any  ;
  description?: string | null ;
  price?: any ;
  quantity?: number ;
  colorsAvailaibles?: string | null ;
  sizesAvailaibles?: string | null  ;
  addDate?: any  ;
  categoryName?: string  ;
  populary?: boolean ;
  newProduct?: boolean ;
  particularity?: string | null ;
  image?: any ;
  image2?: any ;
  image3?: any ;
  image4?: any ;
  image5?: any ;
  _links: any ;

}